Update readme and add client example
